pyproject.toml
[project]
name = "portfolio-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "roberttoribio04@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "fastapi (>=0.124.4,<0.125.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
